move rc/sp into memory
