+++ modules/core/core/src/core/log/log_worker.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include "SkrCore/async/async_service.h"
 #include "SkrContainersDef/vector.hpp"
-#include "SkrCore/sp/sp.hpp"
+#include "SkrCore/memory/sp.hpp"
 #include "log_queue.hpp"
 
 namespace skr
