+++ modules/gui/gui/include/SkrGui/render_objects/render_text.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include "SkrGui/framework/render_object/render_box.hpp"
 #include "SkrGui/math/color.hpp"
-#include "SkrCore/sp/sp.hpp"
+#include "SkrCore/memory/sp.hpp"
 #include <variant> // TODO. use skr::variant, here for shit msvc
 
 #ifndef __meta__
