+++ modules/gui/gui/src/backend/paragraph/paragraph.cpp
@@ -1,4 +1,4 @@
-#include "SkrCore/sp/sp.hpp"
+#include "SkrCore/memory/sp.hpp"
 #include "SkrGui/backend/canvas/canvas.hpp"
 #include "SkrGui/framework/painting_context.hpp"
 #include "SkrGui/_private/paragraph.hpp"
