+++ modules/gui/renderer/include/SkrGuiRenderer/resource/skr_image_task.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "SkrCore/sp/sp.hpp"
+#include "SkrCore/memory/sp.hpp"
 #include "SkrGraphics/cgpux.h"
 #include "SkrRT/io/ram_io.hpp"
 #include "SkrRT/io/vram_io.hpp"
