+++ modules/tools/shader_compiler/src/shader/dxc_compiler.cpp
@@ -1,4 +1,4 @@
-#include "SkrCore/sp/sp.hpp"
+#include "SkrCore/memory/sp.hpp"
 #include "SkrProfile/profile.h"
 #include "SkrCore/log.h"
 #include "SkrContainers/stl_string.hpp" //sv::starts_with
